Add Views, Functions, Show Recent Records, etc.
Added Views to the table list so that the top 100 results and desc
functions can be run on views. Added function query to get the
definition of functions. Added a query to get top 100 results from a
table ordered by id desc. Last, added a query to find every table with a
given column name.