Made Populating Quick Select Synchronous
While async queries are nice in that they don't freeze up the entire UI, this isn't always bad when it comes to the quick select search. Fixes issue where typed letters immediately after hitting the key combo to open the quick select would end up in whatever buffer happened to be open instead of in the search input.